Improved keeping user informed on sync progress
Implemented binding to LibrarySyncService to check if there are items currently
syncing or queued to sync. This makes it possible to inform user of background
sync processes.

Created two abstract classes to reduce code duplication and check if
LibrarySyncService has any items queued which are currently being displayed.

Fixed issue with SwipeRefreshLayout from appcompat library, which does not
always show the refresh animation when refresh is set to true.