Implemented using poster image as placeholder
In most detailed views of a music album, tv show, or movie, the same image is
shown as is used in the overview of artists/albums, tv shows, or movies.
Instead of showing a CharacterDrawable as placeholder while the poster is loading
we now show the image from the overview as placeholder.