Implemented shared element transition for TV shows
Added SharedElementCallback to be able to detect if shared element
is still visible. If not, the shared element transition should not
be performed when returning to the previous fragment.
Added the pager tab strip to the fade animation to keep shared element
transition smooth when poster is partly below the toolbar.