Allow specifying HTTPS Hosts manually
This is not exposed in the UI,
but on the manual host configuration screen,
if the address starts with a "http://" or "https://"
this will be stripped and the `isHttps` flag set appropriately.

Additionally, the port can be configured as part of the address as well.
Any port specified as part of the address takes precedence over
the explicit HTTP port box, to allow easy copy-pasting
of an entire URL for a host.

This enables using Kore with Kodi instances available though HTTPS
(e.g. via a reverse proxy).