Universal Subtitles Options Dialog
this commit fixes #375

A class (app\src\main\java\org\xbmc\kore\ui\generic\SubtitlesOptionsDialog.java) was developed that handles the display of a dialog with options related to subtitles (Download subtitles, Sync subtitles, none). This class was used to add the subtitle button to the widget NowPlayingPanel and the RemoteFragment. Furthermore, any developer can use this class in order to implement a subtitle button to any activity.