Reinstante the ability to bookmark addons
Somewhere along the line, the edits made in b0f2adb876d6650b002bb97042491c227abfe53f where lost because the `AddonsActivity` was instantiating `AddonListFragment` instead of `AddonListContainerFragment`.
This PR reinstates the ability to bookmark and access addons, and refactors the way those settings were being stored in the Shared Preferences to use the default Shared Preferences and take into account the specific host we're connected to.