Fix download files to storage for Android 11+
Kore download files to the default external media store directory, so on Android 11+ (API Level 29) it is within the rules of Scoped Storage, so there's no need to ask for permission to write on external storage.
On Android 10 (API Level 29), that should also aply, but there's a glitch, as apparently Android 10 doesn't grant permissions for the default java `File` interface to write on the default media store directory, so it can't create the download directories. To circunvent this, Kore opts out of Scoped Storage and explicitly asks for external storage write permission, like in previous versions.
Bellow that, we always ask for permission.