Integrate playback with Android's MediaSession
This allows Kodi playback to be controlled by various means within Android's ecosystem, for instance by Android Wear or Google Assistant.
The notification was reworked to be integrated with the MediaSession, therefore conforming to the current media notification standards on Android.
Removed the old ConnectionObserversService and IntentNotificationService, as they were confusing and are deprecated now. Currently the listener to pause playback when receiving a phone call isn't working, and should be restored in a future PR.
Reworked the settings removing the option to show the notification, as now this is always shown.
Added 2 more buttons (skip next and skip previous) to the expanded notification.