Redesign the Artist details screen to show its albums
Instead of showing a screen with tabs for the details/albums/songs of an artist, show the details and albums together on the same screen, making it consistent with other details screens.
Remove the pull to refresh music on details screens, as it doesn't make much sense to refresh the whole music library when looking at a particular item.