Simplify MediaPlaybackBar, MediaActionsBar and MediaProgressIndicator
There's no need for these Views to define an interface to post click events, as these should be handled by the views themselves, making them self-contained.
This also simplifies the seting up of these views, as clients no longer need to call `setDefaultClickListeners`, except on the MediaActionsBar, where a explicit `completeSetup` method must be called to supply a FragmentManager.
There's still a need to call `setPlaybackState` to update their state and UI.