Redesign the Remote screen to show Playback, Progress and Media actions (#852)
On current phone displays the remote screen has more than enough space to show more controls, so make it show buttons for playback control, progress info and media actions. These controls were placed in the Now Playing screen which is not immediately accesible, so moving them to the remote makes it more user-friendly.
Also remove the media actions controls from the Now Playing screen, as these are now redundant. The playback and progress info are also redundant, but we choose to keep them there to keep the UI balanced and because these controls are more useful to the user so they should be accessible in as many places as sensible.