Use standard Floating Action Button for starting playback on media info
According to Android's recent trends the FAB should be unique and represent the most common action on a screen. Therefore, using a single and platform specific FAB more closely follows these guidelines.
The other action on the previous FAB - Play Locally - is moved to the Info Actions Buttons, as it is a secondary action.
The Open IMDb action is removed on the Movie info screen, as the button bar should have at most 4 buttons. The link to IMDb will be reinstated in a future commit.