Improve showing lists when there's no connection
When there's no connection to Kodi, disable the Swipe to Refresh, hide the item list and show the empty view with a message that we're trying to connect to Kodi.
This behaviour is overriden if the list doesn't need a connection to show, as is the case of items that are stored in the local database, descendents from AbstractCursorListFragment
Improve Favourites and Addons list, to only fetch their items when there's a connection.