Fix Addon and Kodi files navigation
Navigation on addons was broken, given that relationships between parent/child directories wasn relying on the same rules as in a filesystem, which is not valid in addons.
This correctly sets the relationship, saving it in the FileLocation structure, and using it to do the navigation.
It also simplifies the code, given that it is no longer necessary to keep the root contents available, and it is easier to check if we're at root level.