Improve jank on list/detail transitions
Transitions had a lot of jank that somehow was caused by setting the exit/reenter transition on the leaving fragment (the list). Removing that transition eliminates the jank, with little visual effects.
Also simplified the shared element and entering fragment transitions, as, not existing a previous exit transition, there's no need to specify a delay on these transitions.