Add collapsing toolbar, review Now Playing panel and play FAB
The initial goal of this was: 1) to hide the toolbar in media screens when scrolled, according to Android motion design and 2) to review the Now Playing panel, as it had some issues that needed addressing, namely 1) it was using a library that is old and no longer maintained and 2) it required that the list and media info `fragments` were inside the Now Playing panel layout, so that it could control those fragments when the panel was shown/hidden. This made it non-obvious the structure of the layout when looking at the base layout (`activity_generic_media`) and some issues as detailed bellow.
These 2 goals were interwined, as to add a collapsing toolbar it needs to:
- Add a `CoordinatorLayout` to the root of `activity_generic_media` to control the toolbar
- Add a `AppBarLayout` enclosing the toolbar
- Make the dependent scroll view a child of the `CoordinatorLayout`, immediatly after the toolbar. This was an issue, as the dependent scroll view (the list or the nested scroll view of the info screens) was nested inside the Now Playing panel, which caused numerous issues when trying to collapse the toolbar.

So to solve this the following changes were done:
- Remove the `SlidingUpPanel` library, replacing it with the default Android mechanism for bottom sheets, which relies on just adding a `BottomSheetBehaviour` to an appropriate view. Now the `NowPlayingPanel` is a view, inheriting from LinearLayout, which doesn't contain the list/info fragment. A dependent fragment can be set on the panel, so that when it goes from the hidden to the collapsed state, it changes the dependent view bottom margins, so that it's still possible to scroll till the end of that view with the panel visible.
- Add the `CoordinatorLayout`, the `AppBarLayout` and move the list/info `FragmentContainerView` to the `activity_generic_media` so that 1) the collapsing toolbar worked, 2) the structure of the screen was obvious and 3) the panel worked

But now we had 2 `CoordinatorLayout` (one inside the other to control the FAB), which caused other issues, namely flickering. Furthermore, the FAB stopped working correctly because the collapsing toolbar changed the dimensions of the fragment that contained the fab, and the fab moved when scrolling the collapsing toolbar.

To solve this, the FAB had to be moved to the `activity_generic_media`, which solves the UI problems, but creates a conceptual issue, as the FAB is conceptually owned and should be controlled by the child info fragment. So this solution breaks some encapsulation, having the fab at the activity, but letting the fragment control it, but it was the only way to make the UI work correctly with the collapsing toolbar, hideable panel and moveable FAB.