Fix sharing play intents only working the first time (#915)
When sharing play intents, the `OpenShareActivity` launches the `RemoteActivity` at the end. In Android versions prior to 12, when doing a second share, it doesn't run `OpenShareActivity` and instead directly opens the `RemoteActivity`, so the play intent isn't delivered to Kodi. This doesn't happen on Android 12+, but to fix it we need to set flags when calling `RemoteActivity` `startActivity` so that the `OpenShareActivity` is called in the next sharing intent.