Update minSdk version to 24
Updating the minSdk version allows use of more Java 8 language features (without desugaring need) and the removalof code that dealt with older versions.
Given that api level 24 (Android 7) come out in 2016 and usage of Android versions 5 and 6 (Api levels 21, 22, 23) are bellow 1% and decreasing (as reported by Google Play), this seems an appropriate tradeoff.