Remove connection/disconnection to LibrarySyncService from fragments that don't have sync items
The creation and connection to the LibrarySyncService was being done by all descendants of AbstractInfoFragment, but this only needs to be done by fragments that have sync items. So, for instance the Addons info fragment was creating and connecting to the service even though it didn't have anything to do with it.

This only creates and connects to the service if the fragment signals that it can use it (has sync items).