PluginUrlUtils: support parsing youtube.com/{live,shorts} URLs
YouTube now use the new format for live streams and shorts. The pattern
is https://www.youtube.com/{live,shorts}/<video id>. But other than
pattern change, the video ID is otherwise completely compatible with the
old format.

Since this pattern is similar to https://youtu.be/ format, re-use that
code by re-shuffle the conditions a bit. The change is applied to both
default and Invidious plugins.

Fixes: https://github.com/xbmc/Kore/issues/966