fix: respect user's intention to queue
before this, the music queue would _always_ be cleared, even if the
user choose the "Queue on Kodi" share action. This is hardly what you expect when you press a button
named "queue".

This also brings Kore's behavior more in line with similar
code (e.g. startPlaylistIfNoActivePlayers in LocalMediaFileListFragment).

Some of the language here was editied to be a bit less video-centric, as well.