misc: improve comment
This else is also hit when the intent contains mixed content. Not the
most elegant solution for mixed content, but the original behavior
of this block was clearly: "Don't know what to do? Must be a video I
guess". This retains that spirit, for better or worse.

Other, slightly overkill option is to extend the PLAYLISTID
types (Kodi's API recognizes -1, which Kodi code defines as 'NONE').