feat: enable sharing multiple files to Kore (#1043)
* refactor: move URI extraction logic into dedicated function

The new function should act identically, but is not conflacted with
URL conversion, and avoids a potential duplicate `intent.getData();` call.

* feat: support SEND_MULTIPLE action

* chore: rename variable

the context being used here does not have to be the application
context. Shortening this name leads to more self-documenting
code (such as when context == some Activty, as it sometimes is), and
is much shorter, to boot.

* chore: fix capitalization

* fix: respect user's intention to queue

before this, the music queue would _always_ be cleared, even if the
user choose the "Queue on Kodi" share action. This is hardly what you expect when you press a button
named "queue".

This also brings Kore's behavior more in line with similar
code (e.g. startPlaylistIfNoActivePlayers in LocalMediaFileListFragment).

Some of the language here was editied to be a bit less video-centric, as well.

* perf: remove unneeded negation

I also think throwing around ! in your code makes it harder to read

* chore: shorten comment

* style: typo

* fix: typo in manifest

* fix: only display Kore on succesfull shares

* misc: improve comment

This else is also hit when the intent contains mixed content. Not the
most elegant solution for mixed content, but the original behavior
of this block was clearly: "Don't know what to do? Must be a video I
guess". This retains that spirit, for better or worse.

Other, slightly overkill option is to extend the PLAYLISTID
types (Kodi's API recognizes -1, which Kodi code defines as 'NONE').

* style: always put .get() on its own line

This was the original convention in this file (tho not necessarily in
the codebase), probably to avoid overly long lines.

I'm happy to revert this change, and have all .get() calls not open a newline.